Sync: New types `LeafPath`, `InteriorPath` and support functions
`InteriorPath` is a path to an interior node in an Ethereum hexary trie.  This
is a sequence of 0 to 64 hex digits.  0 digits means the root node, and 64
digits means a leaf node whose path hasn't been converted to `LeafPath` yet.

`LeafPath` is a path to a leaf in an Ethereum hexary trie.  Individually, each
leaf path is a hash, but rather than being the hash of the contents, it's the
hash of the item's address.  Collectively, these hashes have some 256-bit
numerical properties: ordering, intervals and meaningful difference.