Sync: Add `onGetNodeData`, `onNodeData` to `eth/65` protocol handler
These hooks allow new sync code to register to provide reply data or consume
incoming events without a circular import dependency involving `p2pProtocol`.

Without the hooks, the protocol file needs to import functions that consume
incoming network messages so the `p2pProtocol` can call them, and the functions
that produce outgoing network messages need to import the protocol file.

But related producer/consumer function pairs are typically located in the same
file because they are closely related.  For example the producer of
`GetNodeData` and the consumer of `NodeData`.

In this specific case, we also need to break the `requestResponse` relationship
between `GetNodeData` and `NodeData` messages when pipelining.

There are other ways to accomplish this, but this way is most practical, and
it allows different protocol-using modules to coexist easily.  When the hooks
aren't set, default behaviour is fine.