Working on making vm2 async. (Not done yet.)
This code is not even close to being ready for use.

"make test" should still work, though.

What's in here:

  - There's a duplicated async version of the vm2 inner
loop. (Search for UGLY-duplicatedForAsync.) I'm not happy
with how much duplication there is, and maybe there's some
clever macro I could write to make code that works for both
sync and async, but doing macros-on-macros strikes me as a
little bit too much magic for now, especially since I'm
currently struggling with compiler bugs.

  - The asyncness has been propagated upward, to the
point where I can at least run it through some tests. (I
cut the propagation along other paths, though. Search for
FIXME-eventuallyPropagateAsyncFurtherUpward.)

  - I've got a basic test sub-suite that just runs a few
of the EVM-specific tests. (Run "make test-evm".) I'm
wondering whether was an easier way for me to create a
sub-suite of tests. (See FIXME-howDoIMakeASubSuite.)

  - In the implementation of each opcode, we can now
(optionally) specify an asynchronous handler. (Search for
asyncHandlers.)

  - And the thing that's broken because I encountered that
Nim compiler bug is in op_handlers.nim; there are a few
lines of code in there that you can uncomment to see the
bug. What's *supposed* to happen is that if asyncHandlers
is present, its "run" proc gets put into runAsynchronously
in the vmOpHandlersRec structure. But for some reason that
I don't understand, that's not working. Anyway, the idea
is that when the async version of the EVM is running
an opcode that has an asynchronous handler, it should
run that instead of the synchronous handler. That should
allow us to implement asynchronous handlers for the few
opcodes that need them, while still allowing us to have
a fully-synchronous version of the EVM (which might be
useful for performance comparisons, until/unless we
become satisfied that we don't need it).