NOT WORKING YET: New "stateless" mode.
The basic idea is that just *before* we're about to need some account or slot or block header, we prefetch it asynchronously. (e.g. Accounts and slots are fetched by using eth_getProof, then putting the fetched nodes into the database.)

You can start in this mode by running:
    nimbus --sync-mode=stateless --stateless-data-source-url=https://mainnet.infura.io/whatever
(At which point nimbus-eth1 will wait for an eth2 client to feed it blocks to run via newPayload in the Engine API.)

Alternatively, you can run the block with a particular hash by running:
    nimbus statelesslyRun --stateless-data-source-url=https://mainnet.infura.io/whatever --stateless-block-hash=1234ABCetc.
(This will simply fetch the block header corresponding to that hash from the data source, then run it, without bothering to wait for an eth2 client.)

I say "not working yet" because the state roots are still coming out wrong.

Also, block processing times are still much too slow to be practical: on the order of five minutes per block. We'll need to do precalculated witnesses or something like that in order to make this viable.

One more note: I've removed the concurrentAssemblers test that I implemented when I did the basic async EVM stuff; it was useful as a temporary way to exercise the async code, but the test was based on a very hacky idea and it was more trouble than it was worth.