Timestamp-based hard forks, starting with Shanghai
Replaced the ForkToBlockNumber array with a (more complicated) object called ForkTransitionTable that knows the difference between block-based forks, time-based forks, and the special cases like MergeFork and Shanghai (which IIUC needs to be doable via either block number or time, at least temporarily).

Created a type called ForkDeterminationInfo to represent the information needed to determine which hard fork a particular block is on. In the past all we needed was a block number; now we need to pass in both a block number and timestamp (and if mergeForkBlock isn't specified, maybe TD as well), so I created a type for it (though I don't love the name).

Several things are still unresolved. (See various places that say FIXME-Adam.) But I think this is progress.