Remove contentSize and contentCount from history network log
These are quite expensive database operations, especially the
contentSize one, and are not that necessary to log periodically
considering dbSize is already logged and is the more important
one. ContentSize is mostly useful to see the database overhead,
which we could also add through some json rpc debug call if
needed.