Update the beacon and state network content key prefixes
The Portal specifications were adjusted so that the content key
prefixes become unique selectors over all networks.

Fixes issue #1924

SSZ Union is still used, but it could be that in the future we
need to move to our own custom serialization here.

The unused enum value and the fact that we have to add these to
the case of is quite annoying but unless we stop using a case
object seems not avoidable.