Remove unused seed_db and related code
The code is untested (only in local testnet test and got
commented there due to being flawed in its current form) and
undocumented. It is also unused as it is not really required
for getting all the data for history network and beacon network
seeded.

Something like this might have to be done for state network but
it is not certain at all that the approach taken here would be
the way to go.

Removing the unused code for now and will link commit to
existing issue.